refactored and added additional mappings
